# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "description": "Payload of a single API error.\n\nСодержимое поля ``error`` в ответе. Используется FE для\nлокализации и анализа: код (``code``) стабилен и пригоден для\n``switch``-логики, ``message`` — текст для пользователя/лога,\n``fields`` — пер-полевые ошибки валидации.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Fields",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
      "ApiErrorResponse": {
        "description": "Top-level error envelope returned by the API.\n\nВнешняя «обёртка», чтобы у любого ответа об ошибке был один и тот\nже ключ верхнего уровня. Эта схема также прокидывается в OpenAPI\nкак ``responses`` для роутов, чтобы FE-генератор типов знал форму.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
      "CellSchema": {
        "description": "Single notebook cell as exchanged over the API.\n\nАтомарная ячейка: код или markdown. ``updated_at`` ставит сам\nклиент (миллисекунды от эпохи) — сервер его не переписывает.\nЭто и есть основа LWW-синхронизации.",
        "properties": {
          "content": {
            "default": "",
            "maxLength": 262144,
            "title": "Content",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "code",
              "markdown"
            ],
            "title": "Kind",
            "type": "string"
          },
          "updatedAt": {
            "minimum": 0.0,
            "title": "Updatedat",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "kind",
          "updatedAt"
        ],
        "title": "CellSchema",
        "type": "object"
      },
      "CellTombstone": {
        "description": "Tombstone record for a cell deleted on the client.\n\nКлиент шлёт «надгробие», чтобы сервер при merge мог понять: эта\nячейка удалена локально с такой-то меткой времени. Если на сервере\nта же ячейка обновилась *позже* ``deleted_at`` — удаление\nигнорируется (см. :func:`merge_cells`).",
        "properties": {
          "deletedAt": {
            "minimum": 0.0,
            "title": "Deletedat",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "deletedAt"
        ],
        "title": "CellTombstone",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NotebookCreate": {
        "description": "Request body for ``POST /notebooks``.\n\nОтличается от ``NotebookBase`` опциональным ``id``: если клиент\nего не присылает, сервер сгенерирует ``uuid4`` сам. Если присылает —\nсоздание становится идемпотентным (см. ``NotebookService.create``).",
        "properties": {
          "cells": {
            "items": {
              "$ref": "#/components/schemas/CellSchema"
            },
            "maxItems": 500,
            "title": "Cells",
            "type": "array"
          },
          "formatVersion": {
            "default": 1,
            "minimum": 1.0,
            "title": "Formatversion",
            "type": "integer"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "NotebookCreate",
        "type": "object"
      },
      "NotebookListItem": {
        "description": "Lightweight notebook row for list endpoints.\n\nНе содержит ``cells`` — только их количество. Это сильно сокращает\npayload листинга и предотвращает выгрузку больших JSONB в ответ\nна ``GET /notebooks``.",
        "properties": {
          "cellsCount": {
            "title": "Cellscount",
            "type": "integer"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "integer"
          },
          "formatVersion": {
            "title": "Formatversion",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "title",
          "formatVersion",
          "createdAt",
          "updatedAt",
          "cellsCount"
        ],
        "title": "NotebookListItem",
        "type": "object"
      },
      "NotebookListResponse": {
        "description": "Paginated response wrapper for ``GET /notebooks``.\n\nСтандартная «обёртка» с метаданными пагинации. ``total`` считается\nотдельным запросом и нужен фронту для построения нумерации страниц.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NotebookListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "NotebookListResponse",
        "type": "object"
      },
      "NotebookPatch": {
        "description": "Request body for ``PATCH /notebooks/{id}``.\n\nСейчас это «full sync document», а не классический partial PATCH:\nклиент шлёт полный массив активных ``cells`` плюс ``deletedCells``\n(надгробия) — сервер делает LWW-merge. Семантически это ближе к\nPUT; решение оставить как PATCH было сознательным (Шаг 9 PR #29).",
        "properties": {
          "cells": {
            "items": {
              "$ref": "#/components/schemas/CellSchema"
            },
            "maxItems": 500,
            "title": "Cells",
            "type": "array"
          },
          "deletedCells": {
            "items": {
              "$ref": "#/components/schemas/CellTombstone"
            },
            "title": "Deletedcells",
            "type": "array"
          },
          "formatVersion": {
            "default": 1,
            "minimum": 1.0,
            "title": "Formatversion",
            "type": "integer"
          },
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "NotebookPatch",
        "type": "object"
      },
      "NotebookResponse": {
        "description": "Detailed notebook representation returned by single-item endpoints.\n\nИспользуется в ответах ``POST/GET/PATCH /notebooks/{id}``. Времена\nотдаются в миллисекундах от эпохи (UTC) — единый формат с фронтом.",
        "properties": {
          "cells": {
            "items": {
              "$ref": "#/components/schemas/CellSchema"
            },
            "maxItems": 500,
            "title": "Cells",
            "type": "array"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "integer"
          },
          "formatVersion": {
            "default": 1,
            "minimum": 1.0,
            "title": "Formatversion",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ownerId": {
            "format": "uuid",
            "title": "Ownerid",
            "type": "string"
          },
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "title": "Updatedat",
            "type": "integer"
          }
        },
        "required": [
          "title",
          "id",
          "ownerId",
          "createdAt",
          "updatedAt"
        ],
        "title": "NotebookResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "MSD Course",
      "url": "https://github.com/larchanka-training/dmc-1-t2-notebook-api"
    },
    "description": "## JS Notebook API\n\nBackend API for the JS Notebook project. Built on a multi-module architecture: each domain module owns its `controllers`, `services` and `schemas`.\n\n**Documentation:**\n- Swagger UI: [`/api/v1/docs`](/api/v1/docs)\n- ReDoc: [`/api/v1/redoc`](/api/v1/redoc)\n- OpenAPI schema: [`/api/v1/openapi.json`](/api/v1/openapi.json)\n",
    "license": {
      "name": "MIT"
    },
    "title": "JS Notebook API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/notebooks": {
      "get": {
        "description": "``GET /notebooks`` — paginated list of the user's notebooks.\n\nQuery-параметры ``limit``/``offset`` валидируются FastAPI (``ge``/``le``);\n``sort``/``order`` валидируются сервисом по whitelist.\n\nArgs:\n    limit: Размер страницы (1..200, по умолчанию 50).\n    offset: Смещение (≥ 0).\n    sort: Поле сортировки (``updatedAt`` / ``createdAt`` / ``title``).\n    order: Направление (``asc`` / ``desc``).\n    current_user: Авторизованный пользователь.\n    service: DI-инстанс сервиса.\n\nReturns:\n    Страница :class:`NotebookListResponse`.",
        "operationId": "list_notebooks_api_v1_notebooks_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "updatedAt",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "title": "Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List notebooks",
        "tags": [
          "Notebooks"
        ]
      },
      "post": {
        "description": "``POST /notebooks`` — create or idempotently return a notebook.\n\nВозвращает 201 при создании, 200 при идемпотентном повторе,\n409 при идемпотентном повторе с другим содержимым. Подробности —\nв :meth:`NotebookService.create`.\n\nArgs:\n    payload: Тело запроса с ``id`` (опционально), ``title``, ``cells``.\n    response: FastAPI-объект ответа (нужен, чтобы понизить статус\n        с 201 до 200 при идемпотентном попадании).\n    current_user: Авторизованный пользователь.\n    service: DI-инстанс сервиса.\n\nReturns:\n    Созданный или существующий ноутбук.",
        "operationId": "create_notebook_api_v1_notebooks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotebookCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookResponse"
                }
              }
            },
            "description": "Notebook already existed for this owner"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Notebook id already exists with different content"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create notebook",
        "tags": [
          "Notebooks"
        ]
      }
    },
    "/api/v1/notebooks/{notebook_id}": {
      "delete": {
        "description": "``DELETE /notebooks/{id}`` — soft-delete a notebook.\n\nФизически запись остаётся в БД, но получает ``deleted_at``. Все\nпоследующие чтения по этому id будут отдавать 404.\n\nArgs:\n    notebook_id: UUID ноутбука.\n    current_user: Авторизованный пользователь.\n    service: DI-инстанс сервиса.",
        "operationId": "delete_notebook_api_v1_notebooks__notebook_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "notebook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notebook Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete notebook",
        "tags": [
          "Notebooks"
        ]
      },
      "get": {
        "description": "``GET /notebooks/{id}`` — fetch a single notebook by id.\n\nArgs:\n    notebook_id: UUID ноутбука.\n    current_user: Авторизованный пользователь.\n    service: DI-инстанс сервиса.\n\nReturns:\n    :class:`NotebookResponse`.",
        "operationId": "get_notebook_api_v1_notebooks__notebook_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "notebook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notebook Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get notebook",
        "tags": [
          "Notebooks"
        ]
      },
      "patch": {
        "description": "``PATCH /notebooks/{id}`` — apply offline sync document.\n\nТело — full sync document (полный набор ``cells`` + ``deletedCells``).\nСервер делает LWW-merge и возвращает обновлённый ноутбук.\n\nArgs:\n    notebook_id: UUID ноутбука.\n    payload: Sync-документ от клиента.\n    current_user: Авторизованный пользователь.\n    service: DI-инстанс сервиса.\n\nReturns:\n    Обновлённый ноутбук после merge.",
        "operationId": "patch_notebook_api_v1_notebooks__notebook_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "notebook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notebook Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotebookPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotebookResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Missing or invalid access token"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch notebook",
        "tags": [
          "Notebooks"
        ]
      }
    }
  },
...
}